premium/backupTab/backend/requirements.txt
rich>=13.0.0
pyyaml>=6.0
python-dateutil>=2.8.0
croniter>=1.4.1
b2sdk>=1.21.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
boto3>=1.26.0
cryptography>=41.0.0